backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
openai==1.10.0

# Configuration
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Speech Recognition (STT)
faster-whisper==1.0.3

# Text-to-Speech (TTS)
# piper-tts==1.2.0  # Windows installation issue, using fallback

# Utilities
python-dateutil==2.8.2
tiktoken==0.5.2

# WebSocket
websockets==12.0
